requirements.txt
# Backend Requirements

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Better-Auth
better-auth==0.1.0

# Vector Database
qdrant-client==1.7.3

# AI/ML
google-generativeai==0.3.2
langchain==0.1.4
langchain-google-genai==0.0.6

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.26.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
